requirements.txt
torch
numpy
matplotlib
pillow
splitfolders
albumentations
tqdm